requirements.txt
streamlit
pyaudio
requests
python-dotenv
chromadb
sentence-transformers